MFB-1676: move the CCDF and Medicaid families, with a directory per program
Each family holds base.py plus one file per white label. Where a white label
runs several programs in the family, each gets its own directory so it can hold
its own spec and tests: TX has four Medicaid programs, MA has two.

Medicaid tests move with their programs. The federal base tests become
tests/test_base.py; the per-state classes are extracted from the shared
<wl>/pe/tests/test_member.py modules into tests/test_<wl>.py, carrying the
imports and module-level constants they reference.

Verified against main: registry keys and class names identical, and all 474
test classes / 2625 test methods still present by name. 3262 tests pass.
