+++ programs/programs/co/pe/member.py
@@ -1,29 +1,9 @@
 from programs.framework.pe_base import PolicyEngineMembersCalculator
-from programs.programs.federal.pe.member import CommoditySupplementalFoodProgram, Medicaid
+from programs.programs.federal.pe.member import CommoditySupplementalFoodProgram
 from programs.programs.federal.pe.member import Wic
 import programs.framework.pe_dependencies as dependency
 from screener.models import HouseholdMember
-
-
-class CoMedicaid(Medicaid):
-    program_code = "co_medicaid"
-    medicaid_categories = {
-        "NONE": 0,
-        "ADULT": 310,
-        "INFANT": 200,
-        "YOUNG_CHILD": 200,
-        "OLDER_CHILD": 200,
-        "PREGNANT": 310,
-        "YOUNG_ADULT": 310,
-        "PARENT": 310,
-        "SSI_RECIPIENT": 310,
-        "AGED": 170,
-        "DISABLED": 310,
-    }
-    pe_inputs = [
-        *Medicaid.pe_inputs,
-        dependency.household.CoStateCodeDependency,
-    ]
+from programs.programs.cross_white_label.medicaid.base import Medicaid
 
 
 class AidToTheNeedyAndDisabled(PolicyEngineMembersCalculator):
