+++ programs/programs/cross_white_label/medicaid/tx/for_pregnant_women/calculator.py
@@ -1,0 +1,41 @@
+"""TxMedicaidForPregnantWomen."""
+
+from programs.programs.cross_white_label.medicaid.base import Medicaid
+from screener.models import HouseholdMember
+import programs.framework.pe_dependencies as dependency
+
+
+class TxMedicaidForPregnantWomen(Medicaid):
+    """
+    Texas Medicaid for Pregnant Women calculator that uses PolicyEngine's calculated benefit amounts.
+
+    This program provides free health insurance for pregnant women who do not have
+    other health insurance coverage.
+
+    Eligibility requirements:
+    - Must be pregnant
+    - Must not have other health insurance
+    - Income eligibility determined by PolicyEngine
+    """
+
+    program_code = "tx_medicaid_for_pregnant_women"
+
+    pe_inputs = [
+        *Medicaid.pe_inputs,
+        dependency.household.TxStateCodeDependency,
+    ]
+
+    def member_value(self, member: HouseholdMember):
+        """
+        Returns the Medicaid benefit value for pregnant women without other insurance.
+        """
+        # Must be pregnant
+        if not member.pregnant:
+            return 0
+
+        # Must not have other health insurance
+        if not member.has_insurance_types(("none",)):
+            return 0
+
+        # Return PolicyEngine-calculated value
+        return self.get_member_variable(member.id)
